+++ packages/stylesheet-loader/src/__tests__/transformer.js
@@ -77,6 +77,22 @@ describe('transformer', () => {
       }
     });
   });
+
+  it('should be returned the value of css variable', () => {
+    const style = transformer.convert({
+      type: 'rule',
+      selectors: [ ':root' ],
+      declarations: [
+        {
+          type: 'declaration',
+          property: '--color-name',
+          value: 'blue'
+        }
+      ]
+    });
+
+    expect(style.colorName).toEqual('blue');
+  });
 });
 
 function parse(code) {
